temp

+++ host_read_write/src/host_hal.cpp
@@ -45,7 +45,7 @@ uint32_t readReg(xclDeviceHandle& handle, uint32_t addr)
   return value;
 }
 
-void writeReg(xclDeviceHandle& handle, uint32_t addr,uint8_t value) 
+void writeReg(xclDeviceHandle& handle, uint32_t addr,uint32_t value) 
 {
 	printf("Writing to Address 0x%x value=0x%x\n",addr,value);
     xclWrite(handle, XCL_ADDR_KERNEL_CTRL, addr, (void*)(&value), 4);
